+++ js/markdown/shared/Markdown.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
-	import { createEventDispatcher } from "svelte";
 	import { copy, css_units } from "@gradio/utils";
-	import type { CopyData } from "@gradio/utils";
 	import { Copy, Check } from "@gradio/icons";
 	import type { LoadingStatus } from "@gradio/statustracker";
 	import { IconButton, IconButtonWrapper } from "@gradio/atoms";
@@ -9,37 +7,57 @@
 
 	import { MarkdownCode } from "@gradio/markdown-code";
 
-	export let elem_classes: string[] = [];
-	export let visible: boolean | "hidden" = true;
-	export let value: string;
-	export let min_height: number | string | undefined = undefined;
-	export let rtl = false;
-	export let sanitize_html = true;
-	export let line_breaks = false;
-	export let latex_delimiters: {
-		left: string;
-		right: string;
-		display: boolean;
-	}[];
-	export let header_links = false;
-	export let height: number | string | undefined = undefined;
-	export let show_copy_button = false;
-	export let loading_status: LoadingStatus | undefined = undefined;
-	export let theme_mode: ThemeMode;
-	let copied = false;
+	let {
+		elem_classes = [],
+		visible = true,
+		value,
+		min_height = undefined,
+		rtl = false,
+		sanitize_html = true,
+		line_breaks = false,
+		latex_delimiters = [],
+		header_links = false,
+		height = undefined,
+		show_copy_button = false,
+		loading_status = undefined,
+		theme_mode,
+		onchange = () => {},
+		oncopy = (val) => {}
+	}: {
+		elem_classes: string[];
+		visible: boolean | "hidden";
+		value: string;
+		min_height: number | string | undefined;
+		rtl: boolean;
+		sanitize_html: boolean;
+		line_breaks: boolean;
+		latex_delimiters: {
+			left: string;
+			right: string;
+			display: boolean;
+		}[];
+		header_links: boolean;
+		height: number | string | undefined;
+		show_copy_button: boolean | undefined;
+		loading_status: LoadingStatus | undefined;
+		theme_mode: ThemeMode;
+		onchange: () => void;
+		oncopy: (val: any) => void;
+	} = $props();
+
+	let copied = $state(false);
 	let timer: NodeJS.Timeout;
 
-	const dispatch = createEventDispatcher<{
-		change: undefined;
-		copy: CopyData;
-	}>();
-
-	$: (value, dispatch("change"));
+	$effect(() => {
+		if (value) {
+			onchange();
+		}
+	});
 
 	async function handle_copy(): Promise<void> {
 		if ("clipboard" in navigator) {
 			await navigator.clipboard.writeText(value);
-			dispatch("copy", { value: value });
+			oncopy({ value: value });
 			copy_feedback();
 		}
 	}