+++ js/_spaces-test/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Dependency updates
 
+- @gradio/client@2.0.2
+- @gradio/theme@0.6.0
+
+## 0.0.1
+
+### Dependency updates
+
 - @gradio/client@2.0.1
 
 ## 0.0.1